requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-core==1.38.2
azure-identity==1.25.2
azure-keyvault-secrets==4.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
isodate==0.7.2
msal==1.35.1
msal-extensions==1.3.1
packaging==26.0
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pytesseract==0.3.13
python-multipart==0.0.22
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
